added api description to controller and resolved sme related issues.

+++ src/main/java/com/toqqa/domain/Sme.java
@@ -42,7 +42,7 @@ public class Sme {
 	private Double deliveryRadius;
 	private Double deliveryCharges;
 	private Boolean isRegisterWithGovt;
-	private Double timeOfDelivery;
+	private Long timeOfDelivery;
 	private String regDoc;
 	private String idProof;
 	private String userId;
